Indirect addressing to opcodes, add helper function.
The indirect addressing returns the address, instead of a reference to
the content of the memory. Thus, indirect addressing is better suited
to be included directly in the opcode table, rather than as a memory
access method.

Also, we have added two helper function to set flags as required on
addition/subtraction and other ALU operations.

Fixed definition of certain opcodes.