Change to exact size for types
On different platforms, the fast size may be larger than thought of
size. For example, if 16 bit corresponds to a 64 bit variable, then
program counter might not overflow properly. Thus, we need to change
the define types, and let the program fail to compile if required
types are not present.