Combine opcode and CPU class
This will allow us to keep an accurate cycle count of the various
opcodes. The cycle count depends on the opcode used, and the type of
memory access. Thus, our options was to keep CPU and opcode class
separate, and make the cycle count a global variable, or combine them
together. As a general heuristic to avoid global variable, I will
combine these classes.

Additionally, the implementation of header has been separated into
many file to logically categorize different kind of function.